feat(kv-cache): use fp16 V storage and PV compute

+++ src/ops/kv_cache/append/launch.cu
@@ -93,8 +93,7 @@ void launch_full(const Tensor& k, const Tensor& v, const Tensor& positions, Cach
     kv_cache_append_full_bf16_kernel<Geometry, Metadata><<<fill_grid, Block, 0, stream>>>(
         static_cast<const __nv_bfloat16*>(k.data), static_cast<const __nv_bfloat16*>(v.data),
         static_cast<const std::int32_t*>(positions.data), metadata,
-        static_cast<__nv_bfloat16*>(cache_k.data), static_cast<__nv_bfloat16*>(cache_v.data),
-        tokens);
+        static_cast<__nv_bfloat16*>(cache_k.data), static_cast<__half*>(cache_v.data), tokens);
     CUDA_CHECK(cudaGetLastError());
 }
 
@@ -111,7 +110,7 @@ void launch_paged(const Tensor& k, const Tensor& v, const Tensor& positions, con
     validate_plan(k, plan);
     if (plan.max_count == 0) return;
     auto* cache_k       = static_cast<__nv_bfloat16*>(cache.k_pages.data);
-    auto* cache_v       = static_cast<__nv_bfloat16*>(cache.v_pages.data);
+    auto* cache_v       = static_cast<__half*>(cache.v_pages.data);
     const auto* input_k = static_cast<const __nv_bfloat16*>(k.data);
     const auto* input_v = static_cast<const __nv_bfloat16*>(v.data);
     const auto* pos     = static_cast<const std::int32_t*>(positions.data);
@@ -132,7 +131,7 @@ void launch_cyclic(const Tensor& k, const Tensor& v, const Tensor& positions, co
     validate_plan(k, plan);
     if (plan.max_count == 0) return;
     auto* cache_k       = static_cast<__nv_bfloat16*>(cache.k.data);
-    auto* cache_v       = static_cast<__nv_bfloat16*>(cache.v.data);
+    auto* cache_v       = static_cast<__half*>(cache.v.data);
     const auto* input_k = static_cast<const __nv_bfloat16*>(k.data);
     const auto* input_v = static_cast<const __nv_bfloat16*>(v.data);
     const auto* pos     = static_cast<const std::int32_t*>(positions.data);
